requirements.txt
Django==1.5.1
IPy==0.75
Markdown==2.0.3
MySQL-python==1.2.3
PyYAML==3.10
South==0.8.1
argparse==1.2.1
distribute==0.6.46
django-appconf==0.6
django-markdown==0.1.0
djangorestframework==2.3.6
dnspython==1.11.0
httplib2==0.8
ipaddr==2.1.10
oauth2==1.5.211
pygeoip==0.2.1
python-openid==2.2.5
six==1.3.0
uWSGI==1.9.13
wsgiref==0.1.2
